intan: Drastically simplify timesync algorithm
Since we lazy-load required information from an mmap'ed file, we can
have all information for time synchronization immediately and can make
the algorithm look at all the data instead of file-based slices that end
arbitrarily.
This improves readability of the code, which actually looks extremely
simple now (just as it is!).