Make sure the wheel build always contains dependency data
Depending on how the wheel was built, it either had dependency data
before, or didn't have it. Python packaging is a mess.
The pyproject.toml file is now the canonical source of all iformation on
eldio, and no matter whether the wheel is built via setup.py or Flit, it
will always have comparable information.
Only building via Flit should be used for PyPI uploads though.