intan: Never apply the initial offset twice
In some circumstances the initial offset was applied to data twice when
interpolating, leading to desynchronized time.
We can also assume that there is an unconditional timesync point close
to the end of the sample, so accomodate for that fact as well and
simplify the code a bit to make it easier to reason about by keeping the
sync map intact when passing it to _make_synced_tsvec().